requirements.txt
# Django Core
Django>=4.2.0,<5.0.0
psycopg2-binary>=2.9.0

# API REST
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0

# Multi-Tenancy
django-tenants>=3.5.0

# Permissões
django-guardian>=2.4.0

# CORS (para frontend separado)
django-cors-headers>=4.2.0

# Backup
django-dbbackup>=4.2.0
boto3>=1.28.0

# Tarefas Assíncronas
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0

# Utilitários
django-extensions>=3.2.0

# Rate Limiting
django-ratelimit>=4.1.0

# Payment Gateway
stripe>=7.0.0

# Monitoring
sentry-sdk>=1.32.0
python-json-logger>=2.0.0
psutil>=5.9.0  # Para métricas de sistema (CPU, memória, disco)

# API Documentation
drf-spectacular>=0.27.0

# Reports - PDF Generation
WeasyPrint>=60.0

